chore: extract feel utils
