+++ rules/camunda-cloud/feel.js
@@ -1,9 +1,4 @@
-const { isString } = require('min-dash');
-
-const {
-  is,
-  isAny
-} = require('bpmnlint-utils');
+const { is } = require('bpmnlint-utils');
 
 const { lintExpression } = require('@bpmn-io/feel-lint');
 
@@ -17,20 +12,7 @@ const { ERROR_TYPES } = require('../utils/error-types');
 const { skipInNonExecutableProcess } = require('../utils/rule');
 const { annotateRule } = require('../helper');
 
-// Properties ignored globally
-const IGNORED_PROPERTIES = [
-  'name'
-];
-
-// Properties ignored only for specific element types
-const IGNORED_PROPERTIES_BY_TYPE = {
-  'zeebe:Input': [ 'target' ],
-  'zeebe:Output': [ 'target' ],
-  'zeebe:Header': [ 'key', 'value' ],
-  'zeebe:Property': [ 'name', 'value' ],
-  'zeebe:CalledDecision': [ 'resultVariable' ],
-  'zeebe:Script': [ 'resultVariable' ]
-};
+const { isFeelProperty, findParentNode } = require('./utils/feel');
 
 module.exports = skipInNonExecutableProcess(function() {
   function check(node, reporter) {
@@ -88,26 +70,3 @@ module.exports = skipInNonExecutableProcess(function() {
     check
   });
 });
-
-const isFeelProperty = (node, propertyName, value) => {
-  return !isIgnoredProperty(node, propertyName) && isString(value) && value.startsWith('=');
-};
-
-const isIgnoredProperty = (node, propertyName) => {
-  if (propertyName.startsWith('$') || IGNORED_PROPERTIES.includes(propertyName)) {
-    return true;
-  }
-
-  const nodeType = node.$type;
-  const ignoredForType = IGNORED_PROPERTIES_BY_TYPE[nodeType];
-
-  return ignoredForType && ignoredForType.includes(propertyName);
-};
-
-const findParentNode = node => {
-  while (node && !isAny(node, [ 'bpmn:FlowElement', 'bpmn:FlowElementsContainer' ])) {
-    node = node.$parent;
-  }
-
-  return node;
-};
